% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinationMap.R
\name{bootstrap_rec_map}
\alias{bootstrap_rec_map}
\title{Bootstrap the recombination map}
\usage{
bootstrap_rec_map(
  x,
  intervals,
  nboot = 1000,
  set_negative_values = 0,
  verbose = TRUE
)
}
\arguments{
\item{x}{a 'mareyMap' object.}

\item{intervals}{a data.frame of genomic windows in which the recombination rate will be estimated, with columns 'start'-'end'.}

\item{nboot}{number of bootstraps to estimate the confidence interval.}

\item{set_negative_values}{the value (e.g. 0 or NA) to use for negative values in recombination rates (default = 0).}

\item{verbose}{Whether to print messages and progress bars at each step (default = TRUE).}
}
\value{
a data frame of a recombination map.
}
\description{
Bootstrap the recombination map
}
