% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinationMap.R
\name{bootstrapMareyMap}
\alias{bootstrapMareyMap}
\title{Bootstrap the Marey map}
\usage{
bootstrapMareyMap(x, intervals, nboot = boot, verbose = TRUE)
}
\arguments{
\item{x}{a 'mareyMap' object.}

\item{intervals}{a data.frame of genomic windows in which the recombination rate will be estimated, with columns 'start'-'end'.}

\item{nboot}{number of bootstraps to estimate the confidence interval.}

\item{verbose}{Whether to print messages and progress bars at each step (default = TRUE).}
}
\value{
a data frame of a Marey map with bootstrapped genetic positions.
}
\description{
A generic function to bootstrap a C.I. for a Marey map on given intervals
}
