% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken_stick.R
\name{brokenstick}
\alias{brokenstick}
\title{Plot the broken stick}
\usage{
brokenstick(marey, k = 10, method = "strict", plot = TRUE)
}
\arguments{
\item{marey}{a `comparative_marey_map` object}

\item{k}{the number of segments (default = 10)}

\item{method}{the method to infer the breakpoint of segments, either `strict` to cut a marker position, or `segmented` to interpolate segments breakpoints}

\item{plot}{(logical) whether to plot the broken stick directly or return a data frame (default = `TRUE` will plot the figure)}
}
\value{
a `ggplot` of the broken stick or a data frame, depending on the value of `plot`
}
\description{
Plot a broken stick to compare among a large set of species
the heterogeneity in the distribution of recombination rates along the genome
For a description of the method, see Brazier, T., & Glémin, S. (2022).
Diversity and determinants of recombination landscapes in flowering plants.
PLOS Genetics, 18(8), Article 8. https://doi.org/10.1371/journal.pgen.1010141
}
\details{
Broken Stick model (see White & Hill 2020 for details)
K segments of equal genomic size (Mb) and evaluate the genetic relative size
}
