% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSmoothing.R
\name{calibrate_smoothing}
\alias{calibrate_smoothing}
\title{Automatic calibration of the smoothing parameter}
\usage{
calibrate_smoothing(
  x,
  method = "loess",
  n_resampling = 1000,
  K = 5,
  from = 0.2,
  to = 0.5,
  n_cores = 1,
  degree = 2,
  verbose = TRUE
)
}
\arguments{
\item{x}{a `mareyMap` dataframe, from a slot in a  `marey_map` object}

\item{method}{an interpolation method, either 'loess' or 'splines'.}

\item{n_resampling}{the number of iterations in the cross-validation procedure.}

\item{K}{the number of clusters to subset in K-fold cross-validation.}

\item{from}{the minimum of the range of smoothing values to explore.}

\item{to}{the maximum of the range of smoothing values to explore.}

\item{n_cores}{the number of cores to parallelize.}

\item{degree}{(optional) the degree parameter of the polynomial used in the 'loess' method.}

\item{verbose}{(logical) whether or not to print progress bars}
}
\value{
the best smoothing parameter value.
}
\description{
Automatic calibration of the smoothing parameter
}
