% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSmoothing.R
\name{calibrateSmoothing}
\alias{calibrateSmoothing}
\title{Automatic calibration of the smoothing parameter}
\usage{
calibrateSmoothing(
  x,
  method = "loess",
  nResampling = 1000,
  K = 5,
  from = 0.2,
  to = 0.5,
  nCores = 1,
  degree = 2
)
}
\arguments{
\item{x}{A 'mareyMap' object}

\item{method}{An interpolation method, either 'loess' or 'splines'}

\item{nResampling}{Number of iterations in the cross-validation procedure}

\item{K}{Number of clusters to subset in K-fold cross-validation}

\item{from}{Minimum of the range of smoothing values to explore}

\item{to}{Maximum of the range of smoothing values to explore}

\item{nCores}{Number of cores to parallelize}

\item{degree}{(optional) The degree parameter of the polynomial used in the 'loess' method}
}
\value{
The best smoothing parameter value
}
\description{
Automatic calibration of the smoothing parameter
}
\examples{
calibrateSmoothing(x, "loess")
}
