% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{comparative_interpolation_to_dataframe}
\alias{comparative_interpolation_to_dataframe}
\title{Convert Marey interpolated functions in a `comparative_marey_map` objects to `data.frame`}
\usage{
comparative_interpolation_to_dataframe(x)
}
\arguments{
\item{x}{a `comparative_marey_map` object}
}
\value{
a `data.frame` with `marey_map` interpolation concatenated
}
\description{
Convert Marey interpolated functions in a `comparative_marey_map` objects to `data.frame`
}
