% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{comparative_marey_map}
\alias{comparative_marey_map}
\title{A meta Marey map object for comparative analyses}
\usage{
comparative_marey_map(x = data.frame(), chromosome_length = numeric())
}
\arguments{
\item{x}{a data frame of a Marey map with mandatory columns "set",    "map",    "mkr",    "phys",    "gen".    "vld" is optional and added if not provided.}

\item{chromosome_length}{(optional) a vector of chromosome length. If not provided, max genomic position is assumed to be chromosome end}
}
\value{
a 'comparative_marey_map' object.
}
\description{
This function instantiate a meta marey map object storing mutliple `marey_map` objects
A data frame must be provided with one or many datasets and one or many chromosomes
Marey map functions (e.g. metrics, recombination map estimaion) can be applied in batch
on this object to perform comparative analyses
}
\section{Slots}{

\describe{
\item{\code{data}}{a list of `marey_map` objects}

\item{\code{set}}{a vector of dataset names}

\item{\code{map}}{a vector of chromosome names}
}}

