% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{comparative_recmap_to_dataframe}
\alias{comparative_recmap_to_dataframe}
\title{Convert recombination maps in a `comparative_marey_map` objects to `data.frame`}
\usage{
comparative_recmap_to_dataframe(x)
}
\arguments{
\item{x}{a `comparative_marey_map` object}
}
\value{
a `data.frame` with `marey_map` recombination maps concatenated
}
\description{
Convert recombination maps in a `comparative_marey_map` objects to `data.frame`
}
