% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{comparative_recombination_maps}
\alias{comparative_recombination_maps}
\title{Estimate recombination maps in meta Marey map object}
\usage{
comparative_recombination_maps(x, method = "loess", verbose = TRUE, ...)
}
\arguments{
\item{x}{a `comparative_marey_map` object}

\item{method}{the interpolation method to apply}

\item{verbose}{whether to print messages and progress bar}

\item{...}{additional arguments}
}
\value{
a `comparative_marey_map` object with recombination maps updated
}
\description{
Estimate recombination maps in meta Marey map object
}
