% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvResampling.R
\name{cv_resampling}
\alias{cv_resampling}
\title{K-fold systematic cross-validation of the smoothing parameter}
\usage{
cv_resampling(x, K = 5, smooth = numeric(), method = "loess", degree = 2)
}
\arguments{
\item{x}{a `mareyMap` slot of a `marey_map` object).}

\item{K}{number of clusters to subset in K-fold cross-validation.}

\item{smooth}{the smoothing parameter.}

\item{method}{an interpolation method, either 'loess' or 'splines'.}

\item{degree}{(optional) the degree value of the polynomial function of 'loess' method (default 2).}
}
\value{
the goodness-of-fit criterion (mean squared difference).
}
\description{
K-fold systematic cross-validation of the smoothing parameter
}
