% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvResampling.R
\name{cvResampling}
\alias{cvResampling}
\title{K-fold systematic cross-validation of the smoothing parameter}
\usage{
cvResampling(x, K = 5, smooth = numeric(), method = "loess", degree = 2)
}
\arguments{
\item{x}{a marey map (the 'mareyMap' slot of a mareyMap object).}

\item{K}{number of clusters to subset in K-fold cross-validation.}

\item{smooth}{the smoothing parameter.}

\item{method}{an interpolation method, either 'loess' or 'splines'.}

\item{degree}{(optional) the degree value of the polynomial function of 'loess' method (default 2).}
}
\value{
the goodness-of-fit criterion (mean squared difference).
}
\description{
K-fold systematic cross-validation of the smoothing parameter
}
