% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLoess.R
\name{fitLoess}
\alias{fitLoess}
\title{Fit a LOESS function to the Marey map}
\usage{
fitLoess(x, span = numeric(), degree = 2)
}
\arguments{
\item{x}{a 'mareyMap' data frame.}

\item{span}{the span parameter which controls the degree of smoothing.}

\item{degree}{the degree of the polynomials to be used, normally 1 or 2.}
}
\value{
a LOESS object.
}
\description{
Fit a LOESS function to the Marey map
}
