% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpline.R
\name{fitSpline}
\alias{fitSpline}
\title{Fit a LOESS function to the Marey map}
\usage{
fitSpline(x, spar = numeric())
}
\arguments{
\item{x}{a 'mareyMap' object.}

\item{spar}{smoothing parameter, typically (but not necessarily) in (0,1].}
}
\value{
a smoothed spline object.
}
\description{
Fit a LOESS function to the Marey map
}
