% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip.interval}
\alias{flip.interval}
\title{Flip genetic positions in interval}
\usage{
flip.interval(marey, intervals = data.frame(start = 1000, end = 90000))
}
\arguments{
\item{x}{a 'mareyMap' object.}

\item{intervals}{a data frame of genomic intervals to flip, with physical positions in columns 'start' and 'end'.}
}
\value{
the input 'mareyMap' object with flipped genetic positions.
}
\description{
This function allows to flip the genetic positions in a given genomic interval
to get them in the same order as the reference genome. The intervals are given in a data frame with columns 'start' and 'end'.
Each row is an interval with start and end positions in bp.
You can flip the whole chromosome by giving a single interval with an end equal or larger than chromosome size.
}
