% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip.selection}
\alias{flip.selection}
\title{Flip genetic positions with visual selection of markers}
\usage{
flip.selection(x)
}
\arguments{
\item{x}{a `marey_map` object.}
}
\value{
the input 'mareyMap' object with flipped genetic positions.
}
\description{
This function allows to flip the genetic positions by visually selecting the segment to flip.
}
