% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMareyMap.R
\name{get_marey_map}
\alias{get_marey_map}
\title{Get the Marey map of a `marey_map` object}
\usage{
get_marey_map(x)
}
\arguments{
\item{x}{a `marey_map` object.}
}
\value{
a Marey map without markers `vld` == FALSE.
}
\description{
Get the Marey map slot and remove markers where `vld` == FALSE. Also remove rows with NAs.
}
