% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMareyMap.R
\name{getMareyMap}
\alias{getMareyMap}
\title{Get the Marey map of a mareyMap object}
\usage{
getMareyMap(x)
}
\arguments{
\item{x}{a Marey map object.}
}
\value{
a Marey map without markers `vld` == FALSE.
}
\description{
Get the Marey map slot and remove markers where `vld` == FALSE. Also remove rows with NAs.
}
