% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Calculate the Gini index of a recombination map}
\usage{
gini(x, bootstrap = NULL)
}
\arguments{
\item{x}{a `marey_map` object, with a recombination map already estimated.}

\item{bootstrap}{a numeric value of the number of iterations for a bootstrapped 95\% C.I. Set NULL if no bootstrap.}
}
\value{
The value of the Gini index
}
\description{
The Gini index is a measure of statistical dispersion intended to represent inequality between measures, here recombination rates
A Gini coefficient of 0 reflects perfect equality,
while a Gini coefficient of 1 reflects maximal inequality among values
The Gini index is estimated on recombination rates with the `gini.wtd` function from the `dineq` package.
}
