% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz.R
\name{lorenz}
\alias{lorenz}
\title{Provide a numerical data frame to plot the Lorenz curve}
\usage{
lorenz(x)
}
\arguments{
\item{x}{a Marey map object (the Marey map is used).}
}
\value{
a data frame with relative physical and genetic position ordered.
}
\description{
The Lorenz curve is a graphical representation of the heterogeneity of rates in a dataset, here recombination rates.
It plots the ordered distribution of relative recombination rates as a function of their relative physical size.
A Lorenz curve always starts at (0,0) and ends at (1,1).
The more the curve is close to the diagonal, the less heterogeneous is the distribution of recombination rates.
}
