% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mareyMap.R
\name{marey_map}
\alias{marey_map}
\title{Instantiate a `marey_map` object}
\usage{
marey_map(x = data.frame(), chromosome = NA, chromosome_length = numeric())
}
\arguments{
\item{x}{a data frame of a Marey map with columns "set",    "map",    "mkr",    "phys",    "gen",    "vld". If 'chromosome' names is not provided, this data.frame must contain only a single chromosome.}

\item{chromosome}{the name of the chromosome/map to import from 'x'.}

\item{chromosome_length}{an optional vector of chromosome length in Mb.}
}
\value{
a `marey_map` object.
}
\description{
Instantiate a `marey_map` object
}
