% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask.interval}
\alias{mask.interval}
\title{Mask markers in interval}
\usage{
mask.interval(x, intervals)
}
\arguments{
\item{x}{a data frame of a given interval/selection/chromosome with same columns as a 'mareyMap' data frame}

\item{intervals}{a data frame of genomic intervals to mask, with physical positions in columns 'start' and 'end'.}
}
\value{
the input data frame with masked genetic positions.
}
\description{
Mask genetic markers in a Marey map (set vld = FALSE)
}
