% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean.marey_map}
\alias{mean.marey_map}
\title{Mean of recombination rates}
\usage{
\method{mean}{marey_map}(x, ...)
}
\arguments{
\item{x}{a `marey_map` object.}

\item{...}{arguments passed to the generic summary function.}
}
\value{
numeric, the mean of recombination rates.
}
\description{
Mean of recombination rates
}
