% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean.mareyMap}
\alias{mean.mareyMap}
\title{Mean of recombination rates}
\usage{
mean.mareyMap(x)
}
\arguments{
\item{x}{a Marey map object.}
}
\value{
numeric, the mean of recombination rates.
}
\description{
Mean of recombination rates
}
