% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{median.marey_map}
\alias{median.marey_map}
\title{Median of recombination rates}
\usage{
\method{median}{marey_map}(x, ...)
}
\arguments{
\item{x}{a `marey_map` object.}

\item{...}{arguments passed to the generic summary function.}
}
\value{
numeric, the median of recombination rates.
}
\description{
Median of recombination rates
}
