% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{median.mareyMap}
\alias{median.mareyMap}
\title{Median of recombination rates}
\usage{
\method{median}{mareyMap}(x)
}
\arguments{
\item{x}{a Marey map object.}
}
\value{
numeric, the median of recombination rates.
}
\description{
Median of recombination rates
}
