% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{merge_comparative_marey}
\alias{merge_comparative_marey}
\title{Merge `comparative_marey_map` objects}
\usage{
merge_comparative_marey(x = list())
}
\arguments{
\item{x}{a list of `comparative_marey_map` objects to be merged}
}
\value{
a `comparative_marey_map` object of Marey maps
}
\description{
Merge `comparative_marey_map` objects
}
