% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierSelection.R
\name{outlier_selection}
\alias{outlier_selection}
\title{Remove outlying markers by visual zone selection on the Marey map}
\usage{
outlier_selection(x = marey_map())
}
\arguments{
\item{x}{a `marey_map` object.}
}
\value{
a `marey_map` object.
}
\description{
This function allows to graphically remove/keep markers by zone selection.
}
\details{
Select a zone by left-click around markers to remove/keep, and close the zone by a right-click.

The 'vld' status of the points selected in the graphical windows is set to the opposite:
- 'FALSE' if actual status is 'TRUE'
- 'TRUE' otherwise
}
