% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peripherybias.R
\name{peripherybias}
\alias{peripherybias}
\title{Calculate the periphery-bias ratio as in Brazier and Glémin (2022)}
\usage{
peripherybias(x, proportion_periphery = 0.1, chromosome_arm = "random")
}
\arguments{
\item{x}{a Marey map object, with a recombination map already estimated.}

\item{proportion_periphery}{the proportion of the periphery to keep (default = 0.1)}

\item{chromosome_arm}{which chromosome arm (half of the chromosome) to keep (either 'random', 'left', 'right' or 'both')}
}
\value{
a numeric value of the periphery-bias ratio.
}
\description{
Calculate the periphery-bias ratio as in Brazier and Glémin (2022)
}
