% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.marey_map}
\alias{plot.marey_map}
\title{Plot the Marey maps}
\usage{
\method{plot}{marey_map}(x, return.plot = TRUE, ...)
}
\arguments{
\item{x}{a `marey_map` object.}

\item{return.plot}{TRUE or FALSE, whether to return a plot or a data frame (default = TRUE returns the plot)}

\item{...}{arguments passed to the generic summary function.}
}
\value{
a plot of the Marey maps, or a data frame.
}
\description{
Plot the Marey maps
}
