% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_recombinationMap}
\alias{plot_recombinationMap}
\title{Plot the recombination map}
\usage{
plot_recombinationMap(x)
}
\arguments{
\item{x}{a Marey map object.}
}
\value{
a plot of the recombination map.
}
\description{
Plot the recombination map
}
