% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictGeneticMap.R
\name{predict_genetic_map}
\alias{predict_genetic_map}
\title{Interpolate new genetic positions}
\usage{
predict_genetic_map(x, new)
}
\arguments{
\item{x}{a `marey_map` object, with a recombination map.}

\item{new}{a vector of new physical positions.}
}
\value{
a new `marey_map` object with the new set of markers
}
\description{
Interpolate genetic positions (i.e. cumulative genetic distances)
on a new set of physical positions
and return a new `marey_map` object with the new set of markers (i.e. new genetic and physical distances)
When the `predict_genetic_map` function is used, the standard error of predictions is saved in the 'predict.se' column of the mareyMap slot.
(only for the "loess" regression)
}
