% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictGeneticMap.R
\name{predictGeneticMap}
\alias{predictGeneticMap}
\title{Interpolate new genetic positions}
\usage{
predictGeneticMap(
  phys,
  gen,
  new,
  method = "loess",
  smoothing = 0.3,
  degree = 2
)
}
\arguments{
\item{phys}{a vector of physical positions.}

\item{gen}{a vector of genetic positions.}

\item{new}{a vector of new physical positions.}

\item{method}{an interpolation method, either 'loess' or 'splines'.}

\item{smoothing}{(optional) smoothing parameter, if you do not want to calibrate it.}

\item{degree}{(optional) the degree parameter of the polynomial used in the 'loess' method.}
}
\value{
a list with 'newGen', a vector of new genetic positions for each new marker position and 'se' the standard error of the interpolated result (if relevant for the method).
}
\description{
Interpolate genetic positions (i.e. cumulative gnetic distances)
on a new set of physical positions
}
