% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinationMap.R
\name{recombination_map}
\alias{recombination_map}
\title{Global function estimate the recombination map}
\usage{
recombination_map(
  x,
  method = "loess",
  K = 5,
  boot = 1000,
  n_cores = 1,
  smoothing = numeric(),
  n_resampling = 1000,
  calibration_range = c(0.2, 0.5),
  degree = 2,
  windows = 10^5,
  set_negative_values = 0,
  verbose = TRUE
)
}
\arguments{
\item{x}{a `marey_map` object.}

\item{method}{an interpolation method, either 'loess' or 'splines' (default = 'loess').}

\item{K}{number of clusters to subset in K-fold cross-validation.}

\item{boot}{number of bootstraps to estimate the confidence interval.}

\item{n_cores}{number of cores to parallelize.}

\item{smoothing}{(optional) smoothing parameter, if you do not want to calibrate it.}

\item{n_resampling}{number of iterations in the cross-validation procedure.}

\item{calibration_range}{range of the smoothing parameter space to explore.}

\item{degree}{(optional) the degree parameter of the polynomial used in the 'loess' method.}

\item{windows}{size or coordinates of the windows along the chromosome: either integer in basepairs (default = 10^5) or a data.frame of start/end coordinates.}

\item{set_negative_values}{the value (e.g. 0 or NA) to use for negative values in recombination rates (default = 0).}

\item{verbose}{Whether to print messages and progress bars at each step (default = TRUE).}
}
\value{
a `marey_map` object with the estimated recombination map.
}
\description{
Global function estimate the recombination map
}
