% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinationMap.R
\name{recombinationMap}
\alias{recombinationMap}
\title{Estimate the recombination map}
\usage{
recombinationMap(
  x,
  chromosome = character(),
  method = c("loess", "spline"),
  K = 5,
  boot = 1000,
  nCores = 1,
  smoothing = numeric(),
  nResampling = 1000,
  calibrationRange = c(0.2, 0.5),
  degree = 2,
  windowSize = 10^5
)
}
\arguments{
\item{x}{A 'mareyMap' object}

\item{chromosome}{The name of the chromosome to process}

\item{method}{An interpolation method, either 'loess' or 'splines'}

\item{K}{Number of clusters to subset in K-fold cross-validation}

\item{boot}{Number of bootstrap to estimate a 95% confidence interval}

\item{nCores}{Number of cores to parallelize}

\item{smoothing}{(optional) Smoothing parameter, if you do not want to calibrate it}

\item{nResampling}{Number of iterations in the cross-validation procedure}

\item{calibrationRange}{Range of the smoothing parameter space to explore}

\item{degree}{(optional) The degree parameter of the polynomial used in the 'loess' method}

\item{windowSize}{Size of the windows along the chromosome in basepairs (default = 10^5)}
}
\value{
A 'mareyMap' object with the estimated recombination map
}
\description{
Estimate the recombination map
}
\examples{
recombinationMap(x, chromosome = "1", method = "loess")
}
