% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinationPlot.R
\name{recombinationPlot}
\alias{recombinationPlot}
\title{Plot the Marey maps}
\usage{
recombinationPlot(x)
}
\arguments{
\item{x}{A Marey map object}
}
\value{
A plot of the Marey maps
}
\description{
Plot the Marey maps
}
\examples{
recombinationPlot(x)
}
