% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeChromosome.R
\name{removeChromosome}
\alias{removeChromosome}
\title{Remove a complete chromosome of the dataset (markers set to vld = FALSE)}
\usage{
removeChromosome(x, chromosome = character())
}
\arguments{
\item{x}{A mareyMap object}

\item{chromosome}{A single chromosome name}
}
\value{
A mareyMap object
}
\description{
Remove a complete chromosome of the dataset (markers set to vld = FALSE)
}
\examples{
removeChromosome(x, "1")
}
