% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{subset_comparative_marey}
\alias{subset_comparative_marey}
\title{Subset a `comparative_marey_map` object}
\usage{
subset_comparative_marey(x, subset = c())
}
\arguments{
\item{x}{a `comparative_marey_map` object}

\item{subset}{logical expression indicating `set` and `map` to keep}
}
\value{
a `comparative_marey_map` object of Marey maps
}
\description{
Subset a `comparative_marey_map` object
}
