% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{summary.comparative_marey_map}
\alias{summary.comparative_marey_map}
\title{Summary of a `comparative_marey_map` object}
\usage{
\method{summary}{comparative_marey_map}(x, ...)
}
\arguments{
\item{x}{a `comparative_marey_map` object to summarize}
}
\value{
a summary
}
\description{
Summary of a `comparative_marey_map` object
}
