% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.marey_map}
\alias{summary.marey_map}
\title{Summary of a `marey_map` object}
\usage{
\method{summary}{marey_map}(object, ...)
}
\arguments{
\item{object}{a `marey_map` object.}

\item{...}{arguments passed to the generic summary function.}
}
\value{
a summary of the `marey_map` object.
}
\description{
Summary of a `marey_map` object
}
