% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mareyMap}
\alias{summary.mareyMap}
\title{Summary of a Marey map object}
\usage{
\method{summary}{mareyMap}(x, ...)
}
\arguments{
\item{x}{a Marey map object.}
}
\value{
a summary of the Marey map object.
}
\description{
Summary of a Marey map object
}
