% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{unmask.marker}
\alias{unmask.marker}
\title{Un-Mask markers in interval/position}
\usage{
unmask.marker(x, position)
}
\arguments{
\item{x}{a data frame of a given interval/selection/chromosome with same columns as a 'mareyMap' data frame}

\item{position}{a data frame of genomic intervals/positions to mask, with physical positions in columns 'start' and 'end', or marker positions with columns 'phys' and 'gen'.}
}
\value{
the input data frame with masked genetic positions.
}
\description{
Mask genetic markers in a Marey map (set vld = TRUE)
}
