% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veller.R
\name{veller}
\alias{veller}
\title{Calculate the r-bar statistic from Veller et al. (2019)
r-bar estimate the intra-chromosomal shuffling}
\usage{
veller(x, genome_size = NA, mapping_function = "none")
}
\arguments{
\item{x}{a `marey_map` object.}

\item{genome_size}{The total genome size must be provided to estiate the intra-chromosomal genetic shuffling.}

\item{mapping_function}{The mapping function used to transform genetic distances (either "none", "haldane" or "kosambi")}
}
\value{
a numeric value of r-bar.
}
\description{
Calculate the r-bar statistic from Veller et al. (2019)
r-bar estimate the intra-chromosomal shuffling
}
