% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{vld.marker}
\alias{vld.marker}
\title{Change the state 'vld' of markers in interval/position}
\usage{
vld.marker(x, position, vld = FALSE)
}
\arguments{
\item{x}{a data frame of a given interval/selection/chromosome with same columns as a 'mareyMap' data frame}

\item{position}{a data frame of genomic intervals/positions to mask, with physical positions in columns 'start' and 'end', or marker positions with columns 'phys' and 'gen'.}

\item{vld}{the new state for 'vld'.}
}
\value{
the input data frame with new state of 'vld' at markers positions.
}
\description{
Change 'vld' for genetic markers in a Marey map (set vld = FALSE)
}
