% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.mean.R
\name{weighted.mean.mareyMap}
\alias{weighted.mean.mareyMap}
\title{Weighted Mean of recombination rates}
\usage{
\method{weighted.mean}{mareyMap}(x, ...)
}
\arguments{
\item{x}{a Marey map object.}

\item{...}{arguments passed to the generic summary function.}
}
\value{
numeric, the weighted mean of recombination rates (weighted by windows size).
}
\description{
Weighted Mean of recombination rates
}
