% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken_stick.R
\name{brokenstick_one_map}
\alias{brokenstick_one_map}
\title{Compute the broken stick model for a single Marey map}
\usage{
brokenstick_one_map(marey, k = 10, method = "strict")
}
\description{
Estimate the proportions of a broken stick model for a `comparative_marey_map` object
i.e. proportion of relative genetic length (cM) in k segments of equal genomic size (bp) along the chromosome
}
