% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSmoothing.R
\name{calibrateSmoothing}
\alias{calibrateSmoothing}
\title{Automatic calibration of the smoothing parameter}
\usage{
calibrateSmoothing(
  x,
  method = "loess",
  nResampling = 1000,
  K = 5,
  from = 0.2,
  to = 0.5,
  nCores = 1,
  degree = 2
)
}
\arguments{
\item{x}{a Marey map dataframe.}

\item{method}{an interpolation method, either 'loess' or 'splines'.}

\item{nResampling}{the number of iterations in the cross-validation procedure.}

\item{K}{the number of clusters to subset in K-fold cross-validation.}

\item{from}{the minimum of the range of smoothing values to explore.}

\item{to}{the maximum of the range of smoothing values to explore.}

\item{nCores}{the number of cores to parallelize.}

\item{degree}{(optional) the degree parameter of the polynomial used in the 'loess' method.}
}
\value{
the best smoothing parameter value.
}
\description{
Automatic calibration of the smoothing parameter
}
