% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvResampling.R
\name{cvResampling}
\alias{cvResampling}
\title{K-fold systematic cross-validation of the smoothing parameter}
\usage{
cvResampling(x, K = 5, smooth = numeric(), method = "loess", degree = 2)
}
\arguments{
\item{x}{A marey map (the 'mareyMap' slot of a mareyMap object)}

\item{K}{Number of clusters to subset in K-fold cross-validation}

\item{smooth}{Smoothing parameter}

\item{method}{An interpolation method, either 'loess' or 'splines'}

\item{degree}{(optional) Degree value of the polynomial function of 'loess' method (default 2)}
}
\value{
Goodness-of-fit criterion (mean squared difference)
}
\description{
K-fold systematic cross-validation of the smoothing parameter
}
\examples{
cvResampling(x)
}
