% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mareyMap.R
\name{mareyMap}
\alias{mareyMap}
\title{A Marey map object}
\usage{
mareyMap(x = data.frame(), chromosomeLength = numeric())
}
\arguments{
\item{x}{a data frame of a Marey map with columns "set",    "map",    "mkr",    "phys",    "gen",    "vld".}

\item{chromosomeLength}{an optional vector of chromosome length in Mb.}
}
\value{
a 'mareyMap' object.
}
\description{
A Marey map object
}
