% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierSelection.R
\name{outlierSelection}
\alias{outlierSelection}
\title{Remove outlying markers by visual zone selection on the Marey map}
\usage{
outlierSelection(x = mareyMap(), chromosome = character())
}
\arguments{
\item{x}{a mareyMap object.}

\item{chromosome}{a single chromosome name to display.}
}
\value{
a mareyMap object.
}
\description{
Remove outlying markers by visual zone selection on the Marey map
}
