% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mareyMap}
\alias{plot.mareyMap}
\title{Plot the Marey maps}
\usage{
\method{plot}{mareyMap}(x)
}
\arguments{
\item{x}{a Marey map object.}
}
\value{
a plot of the Marey maps.
}
\description{
Plot the Marey maps
}
