% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mareyMap}
\alias{plot.mareyMap}
\title{Plot the Marey maps}
\usage{
\method{plot}{mareyMap}(x, return.plot = TRUE)
}
\arguments{
\item{x}{a Marey map object.}

\item{return.plot}{TRUE or FALSE, whether to return a plot or a data frame (default = TRUE returns the plot)}
}
\value{
a plot of the Marey maps, or a data frame.
}
\description{
Plot the Marey maps
}
