% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_marey_map.R
\name{plot_comparative_marey}
\alias{plot_comparative_marey}
\title{Plot comparative Marey maps}
\usage{
plot_comparative_marey(x, group = "set + map")
}
\arguments{
\item{x}{a `comparative_marey_map` object}

\item{group}{the grouping factor in `ggplot`, either `set`, `map` or `set + map`}

\item{...}{arguments passed to the generic summary function.}
}
\value{
a `ggplot2` object of Marey maps
}
\description{
Plot comparative Marey maps
}
