% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_recombinationMap}
\alias{plot_recombinationMap}
\title{Plot the recombination map}
\usage{
plot_recombinationMap(x, return.plot = TRUE)
}
\arguments{
\item{x}{a Marey map object.}

\item{return.plot}{TRUE or FALSE, whether to return a plot or a data frame (default = TRUE returns the plot)}
}
\value{
a plot of the recombination map.
}
\description{
Plot the recombination map
}
