% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictGeneticMap.R
\name{predictGeneticMap}
\alias{predictGeneticMap}
\title{Interpolate new genetic positions}
\usage{
predictGeneticMap(x, new)
}
\arguments{
\item{x}{a 'mareyMap' object, with a recombination map.}

\item{new}{a vector of new physical positions.}
}
\value{
a new 'mareyMap' object with the new set of markers
}
\description{
Interpolate genetic positions (i.e. cumulative genetic distances)
on a new set of physical positions
and return a new 'mareyMap' object with the new set of markers (i.e. new genetic and physical distances)
When the 'predictGeneticMap' function is used, the standard error of predictions is saved in the 'predcit.se' column of the mareyMap slot.
(only for the "loess" regression)
}
