% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mareyMap}
\alias{summary.mareyMap}
\title{Summary of a Marey map object}
\usage{
summary.mareyMap(object, ...)
}
\arguments{
\item{object}{a Marey map object.}

\item{...}{arguments passed to the generic summary function.}
}
\value{
a summary of the Marey map object.
}
\description{
Summary of a Marey map object
}
