% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{variance}
\alias{variance}
\title{Variance of recombination rates}
\usage{
variance(x)
}
\arguments{
\item{x}{a Marey map object.}
}
\value{
numeric, the variance of recombination rates.
}
\description{
Variance of recombination rates
}
